Make logs more clear
