+++ modules/shopinsert.py.DISABLED
